API: move `CDNCacheTagsMixin` out from v2
This mixin is used on v2 and v2, so I'm moving it one level up in the directory structure.
